Split out sidebar and refresh dataset queries
